categorise api routes
