+++ app/router/__init__.py
@@ -1,0 +1,8 @@
+"""
+FILE: __init__.py
+DESCRIPTION: Create a base for all APIs
+AUTHOR: Nuttaphat Arunoprayoch
+DATE: 04-Aril-2020
+"""
+from .v1 import router_api_v1
+from .v2 import router_api_v2
