+++ app/router/v1/__init__.py
@@ -1,0 +1,9 @@
+"""
+FILE: __init__.py
+DESCRIPTION: Create a base for API v1 routers
+AUTHOR: Nuttaphat Arunoprayoch
+DATE: 04-Aril-2020
+"""
+from fastapi import APIRouter
+
+v1 = APIRouter()
